+++ bootstrap/http/start.php
@@ -1,7 +1,7 @@
 <?php
-use Opulence\Applications\Environments\Environment;
 use Opulence\Bootstrappers\ApplicationBinder;
 use Opulence\Bootstrappers\Caching\ICache;
+use Opulence\Environments\Environment;
 use Opulence\Framework\Http\Kernel;
 use Opulence\Http\Requests\Request;
 use Opulence\Routing\Router;
@@ -47,7 +47,7 @@ $application = require_once __DIR__ . "/../../config/application.php";
 $applicationBinder->bindToApplication(
     require_once __DIR__ . "/../../config/http/bootstrappers.php",
     false,
-    $application->getEnvironment()->getName() == Environment::PRODUCTION,
+    $environment->getName() == Environment::PRODUCTION,
     $paths["tmp.framework.http"] . "/" . ICache::DEFAULT_CACHED_REGISTRY_FILE_NAME
 );
 
@@ -56,7 +56,7 @@ $applicationBinder->bindToApplication(
  * Let's get started
  * ----------------------------------------------------------
  */
-$application->start(function () use ($application, $container, $exceptionHandler, $exceptionRenderer) {
+$application->start(function () use ($container, $exceptionHandler, $exceptionRenderer) {
     /**
      * ----------------------------------------------------------
      * Handle the request