+++ src/Project/Bootstrappers/Http/Routing/RouterBootstrapper.php
@@ -1,7 +1,7 @@
 <?php
 namespace Project\Bootstrappers\Http\Routing;
 
-use Opulence\Applications\Environments\Environment;
+use Opulence\Environments\Environment;
 use Opulence\Framework\Bootstrappers\Http\Routing\RouterBootstrapper as BaseBootstrapper;
 use Opulence\Routing\Router;
 use Opulence\Routing\Routes\Caching\ICache;