+++ tests/src/Project/Console/ApplicationTestCase.php
@@ -16,7 +16,7 @@ class ApplicationTestCase extends BaseTestCase
     public function setUp()
     {
         $paths = require __DIR__ . "/../../../../config/paths.php";
-        $environment = require __DIR__ . "/../../../../config/environment.php";
+        $this->environment = require __DIR__ . "/../../../../config/environment.php";
         $this->application = require __DIR__ . "/../../../../config/application.php";
         /** @var IContainer $container */
         $this->container = $container;